+++ paint.h
@@ -3,6 +3,6 @@
 #include "Globals.h"
 #include <string>
 
-RECT createRECT(LONG left, LONG top, LONG width, LONG height);
-void drawText(HDC* hdc, std::string txt, RECT rect, int flag, COLORREF bgcol);
+RECT createRECT(int left, int top, int width, int height);
+void drawText(HDC* hdc, std::string txt, RECT rect, unsigned int flag, COLORREF bgcol);
 void pnt(HWND hWnd, RECT* region);
